as fotos dos pokemons agora aparecem normalmente
