+++ pokedex/src/Components/CardPokemon/CardPokemon.js
@@ -1,15 +1,35 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { Button } from '../Button/Button';
 import { CardContainer } from './styled';
 import { goToDetails } from '../../Router/Coordinator';
 import { useHistory } from 'react-router-dom';
+import axios from 'axios'
 
 export function CardPokemon(props) {
   const history = useHistory();
+  const [pokemonPhoto, setPokemonPhoto] = useState();
+
+  useEffect(() => {
+    getPokemonPhoto();
+  }, []);
+
+  const getPokemonPhoto = (pokemon) => {
+    axios
+      .get(`https://pokeapi.co/api/v2/pokemon/${props.name}`)
+      .then((resposta) => {
+        setPokemonPhoto(resposta.data.sprites.front_default);
+      })
+      .catch((err) => {
+        console.log(err.message);
+      });
+  };
+
+  console.log(pokemonPhoto)
 
   return (
     <CardContainer backgroundImage={props.backgroundImage}>
-      <img src={props.url} />
+      <img src={pokemonPhoto} />
+      <h3>{props.name}</h3>
       <div>
         <Button
           text={props.button1}
