+++ pokedex/src/Pages/HomePage/HomePage.js
@@ -3,16 +3,17 @@ import { Header } from '../../Components/Header/Header';
 import { useHistory } from 'react-router-dom';
 import { goToPokedex } from '../../Router/Coordinator';
 import { CardPokemon } from '../../Components/CardPokemon/CardPokemon';
-import { HomeContainer } from './styled';
+import { HomeContainer, PokemonsContainer } from './styled';
 import axios from 'axios';
 
 export function HomePage() {
   const [pokemons, setPokemons] = useState([]);
-  const [pokemonPhoto, setPokemonPhoto] = useState();
   const history = useHistory();
+
   useEffect(() => {
     getPokemons();
   }, []);
+
   const getPokemons = () => {
     axios
       .get('https://pokeapi.co/api/v2/pokemon/')
@@ -23,23 +24,11 @@ export function HomePage() {
         console.log(err.message);
       });
   };
-  const getPokemonPhoto = (pokemon) => {
-    axios
-      .get(`https://pokeapi.co/api/v2/pokemon/${pokemon.name}`)
-      .then((resposta) => {
-        setPokemonPhoto(resposta.data.sprites.front_default);
-        console.log(pokemonPhoto);
-        return pokemonPhoto;
-      })
-      .catch((err) => {
-        console.log(err.message);
-      });
-  };
 
   const renderPokemons = pokemons.map((pokemon) => {
-    return <CardPokemon button1="add pokedex" url={pokemonPhoto} />;
+    return <CardPokemon button1="add pokedex" name={pokemon.name} />;
   });
-
+  console.log(pokemons)
   return (
     <HomeContainer>
       <Header
@@ -47,7 +36,9 @@ export function HomePage() {
         title="Lista de Pokemons"
         onClick={() => goToPokedex(history)}
       />
-      {renderPokemons}
+      <PokemonsContainer>
+        {renderPokemons}
+      </PokemonsContainer>
     </HomeContainer>
   );
 }
