+++ pokedex/src/Pages/HomePage/styled.js
@@ -6,3 +6,10 @@ export const HomeContainer = styled.div`
   height: 100vh;
   width: 100vw;
 `;
+
+export const PokemonsContainer = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  margin: 20px;
+  justify-content: center;
+`
